# back-end/swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "REST API",
    "description": "Documentation automatically generated by the <b>swagger-autogen</b> module."
  },
  "host": "https://vicara-storage-drive.herokuapp.com",
  "basePath": "/",
...
  "schemes": [
    "http",
    "https"
  ],
  "consumes": [
    "application/json"
  ],
  "produces": [
    "application/json"
  ],
  "paths": {
...
    "/api/folder/upload": {
      "post": {
        "tags": [
          "Folder"
        ],
        "description": "Endpoint used for uploading a folder.",
        "produces": [],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "folderName": {
                  "type": "string",
                  "example": "any"
                },
                "users": {
                  "type": "string",
                  "example": "any"
                },
                "users\r": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/folder/get/{id}": {
      "get": {
        "tags": [
          "Folder"
        ],
        "description": "Endpoint used for viewing folders.",
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          }
        }
      }
    },
    "/api/folder/favourite/{folderIDuserID}": {
      "patch": {
        "tags": [
          "Folder"
        ],
        "description": "Endpoint used for adding folder to favourites.",
        "produces": [],
        "parameters": [
          {
            "name": "folderIDuserID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/folder/rename/{folderIDnewNameUserID}": {
      "patch": {
        "tags": [
          "Folder"
        ],
        "description": "Endpoint used for renaming a folder.",
        "produces": [],
        "parameters": [
          {
            "name": "folderIDnewNameUserID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/folder/addUser/{accessFolderIDmailAddedUserID}": {
      "patch": {
        "tags": [
          "Folder"
        ],
        "description": "",
        "produces": [],
        "parameters": [
          {
            "name": "accessFolderIDmailAddedUserID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/folder/delete/{folderIDuserID}": {
      "delete": {
        "tags": [
          "Folder"
        ],
        "description": "Endpoint used for deleting a folder.",
        "produces": [],
        "parameters": [
          {
            "name": "folderIDuserID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/folder/uploadFile": {
      "post": {
        "tags": [
          "Folder"
        ],
        "description": "Endpoint used for uploading a file in folder.",
        "produces": [],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "folderID": {
                  "type": "string",
                  "example": "any"
                },
                "description": {
                  "type": "string",
                  "example": "any"
                },
                "users": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/folder/getFiles/{dets}": {
      "get": {
        "tags": [
          "Folder"
        ],
        "description": "Endpoint used for viewing files inside Folder.",
        "produces": [],
        "parameters": [
          {
            "name": "dets",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/folder/deletefile/{dets}": {
      "delete": {
        "tags": [
          "Folder"
        ],
        "description": "Endpoint used for deleting files in folders.",
        "produces": [],
        "parameters": [
          {
            "name": "dets",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/file/upload": {
      "post": {
        "tags": [
          "File"
        ],
        "description": "Endpoint used for uploading a file.",
        "produces": [],
        "parameters": [
          {
            "name": "obj",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "description": {
                  "type": "string",
                  "example": "any"
                },
                "users": {
                  "type": "string",
                  "example": "any"
                },
                "users\r": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/file/get/{id}": {
      "get": {
        "tags": [
          "File"
        ],
        "description": "Endpoint used for viewing File.",
        "produces": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/file/getPresignedUrl/{fileNameUserID}": {
      "get": {
        "tags": [
          "File"
        ],
        "description": "Endpoint used for viewing files from AWS bucket ( presigned url ).",
        "produces": [],
        "parameters": [
          {
            "name": "fileNameUserID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/file/rename/{fileIDnewNameUserID}": {
      "patch": {
        "tags": [
          "File"
        ],
        "description": "Endpoint used for renaming a file.",
        "produces": [],
        "parameters": [
          {
            "name": "fileIDnewNameUserID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/file/addUser/{accessFileIDmailAddedUserID}": {
      "patch": {
        "tags": [
          "File"
        ],
        "description": "Endpoint used for sharing files.",
        "produces": [],
        "parameters": [
          {
            "name": "accessFileIDmailAddedUserID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/api/file/favourite/{fileIDuserID}": {
      "patch": {
        "tags": [
          "File"
        ],
        "description": "Endpoint used for adding file to favourites.",
        "produces": [],
        "parameters": [
          {
            "name": "fileIDuserID",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Bad Request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
...
  },
...
}